refactor(clp-tdl-package): Move `clp_binary_path` and `s3_credential_env` into a shared `task::utils` module.
