+++ components/clp-tdl-package/src/task/mod.rs
@@ -1,3 +1,4 @@
 //! The task implementations this package registers with Spider.
 
 pub mod compression;
+pub mod utils;
